Remove 'curses' gem, replace with 'ncurses-ruby'.
Now that 'curses' isn't bundled with ruby, there's no reason to prefer it over
the more functional option.

Because ncurses-ruby exposes newterm(), I was able to remove a LOT of horrible
IPC code.  Previously, we created a socketpair, forked, reopen'd STDIN and
STDOUT, and set up communication channels just because I couldn't tell curses
what terminal to use (e.g., /dev/tty, not STDIN and STDOUT).  All of that stuff
is gone, and with it, much of the complexity of listpager.