Long-sleep mode
This change gets rid of a separate select on stdin + a nonblocking
Ncurses.getch.  The problem was, we were OK blocking in the stdin
select, but because we want to be responsive to keypresses, we had
to keep the select timeout low so that getch didn't get starved.

Now, we have a single select that waits for stdin, the tty, and a
now-required self_pipe.  When we are notified of data on the tty,
we call non-blocking getch until its exhausted.

With this scheme, we no longer got window resizes, because we were
blocking in select.  Now we trap on WINCH, and write a character
to self_pipe.  The select picks that up and handles it in the normal
flow of the main loop.

Commit also finally pulls out classes into their own file.  At 500+
lines, it was time to give up on the idea that this could be a
self-contained program, especially as we still don't process any
command-line arguments, and the list of working command-mode
commands are yet to be filled out.