Implement `%lock/%unlock`.
When a client sends `%lock`, and receives a `%lock` ack back, listpager will no longer send
user-generated events to stdout, but will still respond to the client.  This lets the client
interrogate listpager about its state in a predictable, blocking manner, e.g.:

  > %lock
  < %lock
  > %get-title
  < %title-is "Something"
     # sleep 5
  > %unlock
  < %unlock
     # now here come all the events that were held:
  < %selected-is ...
  < %selected-is ...

Both client and server commands are now `%`-prefixed, this is to allow list/range interrogation
in the future.